--- response-items/freetext-component/.storybook/addons.js
@@ -1,3 +1,0 @@
-import '@storybook/addon-actions/register';
-import '@storybook/addon-links/register';
-import "@storybook/addon-knobs/register";
