+++ response-items/freetext/.storybook/addons.js
@@ -1,0 +1,3 @@
+import "@storybook/addon-actions/register";
+import "@storybook/addon-links/register";
+import "@storybook/addon-knobs/register";
