+++ response-items/freetext/src/Component.stats.js
@@ -6,21 +6,21 @@ const fixedVal = 3;
 const fixed = (fn, ...args) => parseFloat(fn(...args).toFixed(fixedVal));
 
 const stats = (_, results) => {
-  const wordCounts = results.map(x => x.text.split(/\s+/).length);
+  const wordCounts = results.map((x) => x.text.split(/\s+/).length);
   return {
     visualizations: [
       {
         name: "Word Cloud",
-        component: <Visualization values={results} />
-      }
+        component: <Visualization values={results} />,
+      },
     ],
     stats: {
       ["Word Count - Min, Max"]: `${fixed(Math.min, ...wordCounts)}, ${fixed(
         Math.max,
         ...wordCounts
       )}`,
-      ["Word Count - Mean"]: fixed(math.mean, wordCounts)
-    }
+      ["Word Count - Mean"]: fixed(math.mean, wordCounts),
+    },
   };
 };
